requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.4
aiohttp-jinja2==1.6
aiosignal==1.4.0
alembic==1.14.1
arrow==1.4.0
attrs==25.4.0
bcrypt==4.3.0
cffi==2.0.0
click==8.1.7
cryptography==43.0.1
frozenlist==1.8.0
greenlet==3.3.1
idna==3.11
Jinja2==3.0.3
jinja2-time==0.2.0
loguru==0.7.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycares==5.0.1
pycparser==3.0
Pygments==2.19.2
pyjabber==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2022.1
PyYAML==6.0.3
rich==14.3.3
singletonify==0.2.4
six==1.17.0
slixmpp_multiplatform==1.12.0
spade==4.1.2
SQLAlchemy==2.0.40
timeago==1.0.16
typing_extensions==4.15.0
tzdata==2025.3
uvloop==0.21.0
yarl==1.22.0
